pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "Friday"
version = "0.1"
dependencies = ["gitpython", "openai-whisper", "speechrecognition", "sounddevice", "numpy", "pydub"]

[tool.maturin]
python-source = "python"
module-name = "Friday.tts"

[project.urls]
GitHub = "https://github.com/d33pster/Quinn"

[project.scripts]
friday = "Friday:friday.main"